feat: always skip sudo mode for password change of oauth users, fix admin detection
